# backend/cmd/server/main.go
package main

import (
	"context"
	"fmt"
	"log"
	"net/http"
	"os"
	"path/filepath"
	"strings"

	"github.com/gin-gonic/gin"
	"github.com/jackc/pgx/v5/pgxpool"
	"github.com/joho/godotenv"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"

	_ "opscore/backend/docs" // docs is generated by Swag CLI
	authmiddleware "opscore/backend/internal/auth/interfaces/middleware"
	"opscore/backend/internal/shared/infrastructure/middleware"
)

// @title OpsCore Backend API
// @version 1.0
// @description This is the API documentation for the OpsCore backend service.
// @host localhost:8080
// @BasePath /api/v1
// @securityDefinitions.apikey BearerAuth
// @in header
// @name Authorization
// @description Enter "Bearer " followed by your JWT token. Example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
func main() {
	// Load .env file if it exists
	if err := godotenv.Load(); err != nil {
		// .env file not found or error loading - this is not fatal
		// Environment variables can be set directly in the system
		log.Println("No .env file found or error loading it, using system environment variables")
	} else {
		log.Println("Loaded .env file successfully")
	}

	// --- Database Connection ---
	// Use environment variable for connection string (replace with your actual env var name)
	databaseUrl := os.Getenv("DATABASE_URL")
	if databaseUrl == "" {
		// Provide a default for local development if needed, but env var is better
		databaseUrl = "postgres://opscore_user:opscore_password@db:5432/opscore_db?sslmode=disable"
		fmt.Println("Warning: DATABASE_URL environment variable not set, using default.")
	}

	dbpool, err := pgxpool.New(context.Background(), databaseUrl)
	if err != nil {
		fmt.Fprintf(os.Stderr, "Unable to create connection pool: %v\n", err)
		os.Exit(1) // Exit if DB connection fails
	}
	defer dbpool.Close() // Ensure pool is closed when main function exits

	// Ping the database to verify connection
	err = dbpool.Ping(context.Background())
	if err != nil {
		fmt.Fprintf(os.Stderr, "Unable to ping database: %v\n", err)
		os.Exit(1)
	}
	fmt.Println("Successfully connected to the database.")
	// --- End Database Connection ---

	repoHandler, docHandler, varHandler, execHandler, attachHandler, userHandler, groupHandler, viewHistoryHandler, viewStatsHandler, oauthHandler, gitProviderHandler, authHandler, jwtService, appLogger, err := InitializeAPI(dbpool) // Pass dbpool and handle error
	if err != nil {
		fmt.Fprintf(os.Stderr, "Failed to initialize API dependencies: %v\n", err)
		os.Exit(1)
	}

	// Create auth middleware
	authMiddleware := authmiddleware.AuthMiddleware(jwtService)

	// Use gin.New() instead of gin.Default() to have full control over middleware
	r := gin.New()

	// Apply custom middleware in order
	r.Use(middleware.Logger(appLogger)) // Custom structured logger with request_id
	r.Use(gin.Recovery())               // Panic recovery middleware

	// Allow all origins (for development)
	r.Use(func(c *gin.Context) {
		c.Writer.Header().Set("Access-Control-Allow-Origin", "*")
		c.Writer.Header().Set("Access-Control-Allow-Credentials", "true")
		c.Writer.Header().Set("Access-Control-Allow-Headers", "Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization, accept, origin, Cache-Control, X-Requested-With")
		c.Writer.Header().Set("Access-Control-Allow-Methods", "POST, OPTIONS, GET, PUT, DELETE")

		if c.Request.Method == "OPTIONS" {
			c.AbortWithStatus(204)
			return
		}

		c.Next()
	})

	// Swagger documentation endpoint
	// Adjust the import path above before running swag init
	r.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))

	// API v1 routes group
	v1 := r.Group("/api/v1")
	{
		// Authentication routes (Multi-provider OIDC)
		authGroup := v1.Group("/auth")
		{
			// Public routes - no authentication required
			authGroup.GET("/:provider/login", authHandler.ProviderLogin)
			authGroup.POST("/:provider/callback", authHandler.ProviderCallback)
			authGroup.POST("/refresh", authHandler.RefreshToken)

			// Legacy OAuth routes for repository access (returns token without saving)
			authGroup.POST("/oauth/callback", oauthHandler.HandleCallback)
		}

		// Protected auth routes - require authentication
		authProtected := v1.Group("/auth")
		authProtected.Use(authMiddleware)
		{
			// User info and session management
			authProtected.GET("/me", authHandler.GetMe)
			authProtected.POST("/logout", authHandler.Logout)

			// Identity management
			authProtected.GET("/identities", authHandler.GetIdentities)
			authProtected.DELETE("/identities/:id", authHandler.UnlinkIdentity)

			// New OAuth routes for Git provider connections (saves token to database)
			authProtected.POST("/oauth/connect", oauthHandler.HandleCallbackWithSave)
			authProtected.GET("/oauth/connections", oauthHandler.ListConnections)
			authProtected.GET("/oauth/connections/:connectionId/repositories", oauthHandler.ListRepositoriesByConnection)
			authProtected.GET("/oauth/connections/:connectionId/repositories/:owner/:repo/contents", oauthHandler.GetRepositoryContents)
			authProtected.GET("/oauth/connections/:connectionId/repositories/:owner/:repo/files/*filePath", oauthHandler.GetRepositoryFileContent)
			authProtected.DELETE("/oauth/connections/:provider", oauthHandler.DisconnectProvider)
		}

		// Git provider routes - access repositories using OAuth tokens (protected)
		gitProviders := v1.Group("/git-providers")
		gitProviders.Use(authMiddleware)
		{
			gitProviders.GET("/:provider/repositories", gitProviderHandler.ListUserRepositories)
		}

		// Repository routes - public routes
		v1.GET("/repositories", repoHandler.ListRepositories)
		v1.GET("/repositories/:repoId", repoHandler.GetRepository)

		// Repository file routes - require authentication (uses OAuth token)
		repoFiles := v1.Group("/repositories")
		repoFiles.Use(authMiddleware)
		{
			repoFiles.GET("/:repoId/files", repoHandler.ListRepositoryFiles)
			repoFiles.GET("/:repoId/contents", repoHandler.GetRepositoryContents)
			repoFiles.GET("/:repoId/files/content", repoHandler.GetFileContents)
			repoFiles.GET("/:repoId/files/history", repoHandler.GetFileHistory)
		}

		// Document routes - public routes
		v1.GET("/documents", docHandler.ListDocuments)
		v1.GET("/documents/:docId", docHandler.GetDocument)

		// Document routes - protected routes (require authentication)
		documentsProtected := v1.Group("/documents")
		documentsProtected.Use(authMiddleware)
		{
			documentsProtected.POST("", docHandler.CreateDocument)
			documentsProtected.POST("/publish", docHandler.PublishDocument)
			documentsProtected.PUT("/:docId", docHandler.UpdateDocument)
			documentsProtected.PATCH("/:docId/metadata", docHandler.UpdateDocumentMetadata)
			documentsProtected.GET("/:docId/versions", docHandler.GetDocumentVersions)
			documentsProtected.GET("/:docId/versions/:version", docHandler.GetDocumentVersion)
			documentsProtected.POST("/:docId/versions/:version/publish", docHandler.PublishDocumentVersion)
			documentsProtected.POST("/:docId/versions/:version/rollback", docHandler.RollbackDocumentVersion)
		}

		// Variable routes
		v1.GET("/documents/:docId/variables", varHandler.GetVariableDefinitions)
		v1.POST("/documents/:docId/validate-variables", varHandler.ValidateVariableValues)

		// Execution record routes
		v1.POST("/execution-records", execHandler.CreateExecutionRecord)
		v1.GET("/execution-records", execHandler.SearchExecutionRecords)
		v1.GET("/execution-records/:id", execHandler.GetExecutionRecord)
		v1.PUT("/execution-records/:id/title", execHandler.UpdateTitle)
		v1.PUT("/execution-records/:id/notes", execHandler.UpdateNotes)
		v1.PUT("/execution-records/:id/access-scope", execHandler.UpdateAccessScope)
		v1.POST("/execution-records/:id/complete", execHandler.Complete)
		v1.POST("/execution-records/:id/fail", execHandler.MarkAsFailed)
		v1.POST("/execution-records/:id/steps", execHandler.AddStep)
		v1.PUT("/execution-records/:id/steps/:stepNumber/notes", execHandler.UpdateStepNotes)
		v1.DELETE("/execution-records/:id", execHandler.DeleteExecutionRecord)

		// Attachment routes
		v1.POST("/execution-records/:id/attachments", attachHandler.UploadAttachment)
		v1.GET("/attachments/:id", attachHandler.GetAttachment)
		v1.GET("/attachments/:id/download", attachHandler.DownloadAttachment)
		v1.GET("/attachments/:id/url", attachHandler.GetAttachmentURL)
		v1.GET("/execution-records/:id/attachments", attachHandler.ListAttachments)
		v1.GET("/execution-records/:id/steps/:stepId/attachments", attachHandler.ListStepAttachments)
		v1.DELETE("/attachments/:id", attachHandler.DeleteAttachment)

		// User routes
		v1.POST("/users", userHandler.CreateUser)
		v1.GET("/users/:userId", userHandler.GetUser)
		v1.GET("/users", userHandler.ListUsers)
		v1.PUT("/users/:userId", userHandler.UpdateUser)
		v1.DELETE("/users/:userId", userHandler.DeleteUser)

		// Group routes
		v1.POST("/groups", groupHandler.CreateGroup)
		v1.GET("/groups/:groupId", groupHandler.GetGroup)
		v1.GET("/groups", groupHandler.ListGroups)
		v1.PUT("/groups/:groupId", groupHandler.UpdateGroup)
		v1.DELETE("/groups/:groupId", groupHandler.DeleteGroup)
		v1.POST("/groups/:groupId/members", groupHandler.AddMember)
		v1.DELETE("/groups/:groupId/members", groupHandler.RemoveMember)
		v1.GET("/users/:userId/groups", groupHandler.GetUserGroups)

		// View history routes
		v1.POST("/documents/:docId/views", viewHistoryHandler.RecordView)
		v1.GET("/users/:userId/view-history", viewHistoryHandler.GetUserViewHistory)
		v1.GET("/documents/:docId/view-history", viewHistoryHandler.GetDocumentViewHistory)

		// View statistics routes
		v1.GET("/documents/:docId/statistics", viewStatsHandler.GetDocumentStatistics)
		v1.GET("/users/:userId/statistics", viewStatsHandler.GetUserStatistics)
		v1.GET("/statistics/popular-documents", viewStatsHandler.GetPopularDocuments)
		v1.GET("/statistics/recent-documents", viewStatsHandler.GetRecentlyViewedDocuments)
	}

	// Static file serving (Frontend)
	r.StaticFile("/", "./frontend/dist/index.html")
	r.Static("/assets", "./frontend/dist/assets")
	r.StaticFile("/favicon.ico", "./frontend/dist/favicon.ico")
	r.StaticFile("/vite.svg", "./frontend/dist/vite.svg")
	r.NoRoute(func(c *gin.Context) {
		if !strings.HasPrefix(c.Request.URL.Path, "/api") {
			filePath := "./frontend/dist" + c.Request.URL.Path
			absPath, _ := filepath.Abs(filePath)
			if _, err := http.Dir(filepath.Dir(absPath)).Open(filepath.Base(absPath)); err != nil {
				c.File("./frontend/dist/index.html")
				return
			}
			c.File(filePath)
		} else {
			c.JSON(http.StatusNotFound, gin.H{"code": "PAGE_NOT_FOUND", "message": "Page not found"})
		}
	})

	port := "8080"
	println("Backend server listening at http://localhost:" + port)
	r.Run(":" + port)
}