# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  opscore_backend_internal_view_history_interfaces_api_schema.ViewHistoryListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/opscore_backend_internal_view_history_interfaces_api_schema.ViewHistoryResponse'
        type: array
      limit:
        example: 50
        type: integer
      offset:
        example: 0
        type: integer
      total_count:
        example: 100
        type: integer
    type: object
...
  opscore_backend_internal_view_statistics_interfaces_api_schema.DocumentStatisticsResponse:
    properties:
      average_view_duration:
        example: 120
        type: integer
      document_id:
        example: a1b2c3d4-e5f6-7890-1234-567890abcdef
        type: string
      last_viewed_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      total_views:
        example: 1234
        type: integer
      unique_viewers:
        example: 567
        type: integer
    type: object
  opscore_backend_internal_view_statistics_interfaces_api_schema.ErrorResponse:
    properties:
      code:
        example: INVALID_REQUEST
        type: string
      details:
        type: string
      message:
        example: Invalid request format
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation for the OpsCore backend service.
  title: OpsCore Backend API
  version: "1.0"
paths:
  /api/documents/{docId}/statistics:
    get:
      description: Retrieves view statistics for a specific document
      parameters:
      - description: Document ID
        in: path
        name: docId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Statistics retrieved successfully
          schema:
            $ref: '#/definitions/opscore_backend_internal_view_statistics_interfaces_api_schema.DocumentStatisticsResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/opscore_backend_internal_view_statistics_interfaces_api_schema.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/opscore_backend_internal_view_statistics_interfaces_api_schema.ErrorResponse'
      summary: Get document statistics
      tags:
      - statistics
  /api/documents/{docId}/view-history:
    get:
      description: Retrieves the view history for a specific document
      parameters:
      - description: Document ID
        in: path
        name: docId
        required: true
        type: string
      - default: 50
        description: Limit
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: View history retrieved successfully
          schema:
            $ref: '#/definitions/opscore_backend_internal_view_history_interfaces_api_schema.ViewHistoryListResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/opscore_backend_internal_view_history_interfaces_api_schema.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/opscore_backend_internal_view_history_interfaces_api_schema.ErrorResponse'
      summary: Get document's view history
      tags:
      - view-history
  /api/documents/{docId}/views:
    post:
      consumes:
      - application/json
      description: Records a view of a document by a user
      parameters:
      - description: Document ID
        in: path
        name: docId
        required: true
        type: string
      - description: User ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/opscore_backend_internal_view_history_interfaces_api_schema.RecordViewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: View recorded successfully
          schema:
            $ref: '#/definitions/opscore_backend_internal_view_history_interfaces_api_schema.ViewHistoryResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/opscore_backend_internal_view_history_interfaces_api_schema.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/opscore_backend_internal_view_history_interfaces_api_schema.ErrorResponse'
      summary: Record a document view
      tags:
      - view-history
  /api/statistics/popular-documents:
    get:
      description: Retrieves the most popular documents based on view count
      parameters:
      - default: 10
        description: Limit
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 30
        description: Days to look back
        in: query
        minimum: 1
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Popular documents retrieved successfully
          schema:
            $ref: '#/definitions/opscore_backend_internal_view_statistics_interfaces_api_schema.PopularDocumentsResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/opscore_backend_internal_view_statistics_interfaces_api_schema.ErrorResponse'
      summary: Get popular documents
      tags:
      - statistics
  /api/statistics/recent-documents:
    get:
      description: Retrieves the most recently viewed documents
      parameters:
      - default: 10
        description: Limit
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Recent documents retrieved successfully
          schema:
            $ref: '#/definitions/opscore_backend_internal_view_statistics_interfaces_api_schema.RecentDocumentsResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/opscore_backend_internal_view_statistics_interfaces_api_schema.ErrorResponse'
      summary: Get recently viewed documents
      tags:
      - statistics
  /api/users/{userId}/statistics:
    get:
      description: Retrieves view statistics for a specific user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Statistics retrieved successfully
          schema:
            $ref: '#/definitions/opscore_backend_internal_view_statistics_interfaces_api_schema.UserStatisticsResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/opscore_backend_internal_view_statistics_interfaces_api_schema.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/opscore_backend_internal_view_statistics_interfaces_api_schema.ErrorResponse'
      summary: Get user statistics
      tags:
      - statistics
  /api/users/{userId}/view-history:
    get:
      description: Retrieves the view history for a specific user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - default: 50
        description: Limit
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: View history retrieved successfully
          schema:
            $ref: '#/definitions/opscore_backend_internal_view_history_interfaces_api_schema.ViewHistoryListResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/opscore_backend_internal_view_history_interfaces_api_schema.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/opscore_backend_internal_view_history_interfaces_api_schema.ErrorResponse'
      summary: Get user's view history
      tags:
      - view-history
  /attachments/{id}:
    delete:
      description: Delete a specific attachment by ID
      parameters:
      - description: Attachment ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Attachment deleted successfully
        "400":
          description: Invalid attachment ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Attachment not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an attachment
      tags:
      - attachments
    get:
      description: Retrieves metadata information about a specific attachment
      parameters:
      - description: Attachment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved attachment metadata
          schema:
            $ref: '#/definitions/opscore_backend_internal_execution_record_interfaces_api_schema.AttachmentResponse'
        "400":
          description: Invalid attachment ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Attachment not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get attachment metadata
      tags:
      - attachments
  /attachments/{id}/download:
    get:
      description: Download the actual file content of an attachment
      parameters:
      - description: Attachment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Attachment file content
          schema:
            type: file
        "400":
          description: Invalid attachment ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Attachment not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Download an attachment file
      tags:
      - attachments
  /attachments/{id}/url:
    get:
      description: Get a presigned URL for downloading an attachment (for S3 storage)
        or download endpoint (for local storage)
      parameters:
      - description: Attachment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Presigned URL or download endpoint
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid attachment ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Attachment not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get signed URL for attachment
      tags:
      - attachments
  /auth/{provider}/callback:
    post:
      consumes:
      - application/json
      description: Exchanges authorization code for JWT token and creates/updates
        user
      parameters:
      - description: Provider name (google, github, gitlab, microsoft)
        in: path
        name: provider
        required: true
        type: string
      - description: Authorization code and state
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth_interfaces_api_handlers.ProviderCallbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth_interfaces_api_handlers.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Handle OIDC provider callback
      tags:
      - auth
  /auth/{provider}/login:
    get:
      description: Returns OIDC provider authorization URL for user to authenticate
      parameters:
      - description: Provider name (google, github, gitlab, microsoft)
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth_interfaces_api_handlers.ProviderLoginResponse'
      summary: Initiate OIDC login
      tags:
      - auth
  /auth/identities:
    get:
      description: Returns all provider identities linked to the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/internal_auth_interfaces_api_handlers.IdentityResponse'
              type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user's linked identities
      tags:
      - auth
  /auth/identities/{id}:
    delete:
      description: Removes a provider identity from the user's account
      parameters:
      - description: Identity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Unlink a provider identity
      tags:
      - auth
  /auth/logout:
    post:
      description: Revokes all refresh tokens for the user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /auth/me:
    get:
      description: Returns the currently authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth_interfaces_api_handlers.UserProfile'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user information
      tags:
      - auth
  /auth/oauth/callback:
    post:
      consumes:
      - application/json
      description: Exchanges OAuth authorization code for access token
      parameters:
      - description: OAuth callback request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/opscore_backend_internal_oauth_domain.OAuthCallbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/opscore_backend_internal_oauth_domain.OAuthTokenResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: OAuth callback endpoint
      tags:
      - OAuth
  /auth/oauth/connect:
    post:
      consumes:
      - application/json
      description: Exchanges OAuth authorization code for access token and saves to
        database
      parameters:
      - description: OAuth callback request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/opscore_backend_internal_oauth_domain.OAuthCallbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_oauth_interfaces_api_handlers.OAuthConnectionResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: OAuth callback endpoint with token storage
      tags:
      - OAuth
  /auth/oauth/connections:
    get:
      description: List all OAuth connections for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List OAuth connections
      tags:
      - OAuth
  /auth/oauth/connections/{connectionId}/repositories:
    get:
      description: List all repositories accessible via a specific OAuth connection
      parameters:
      - description: OAuth connection ID
        in: path
        name: connectionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid connection ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List repositories by OAuth connection
      tags:
      - OAuth
  /auth/oauth/connections/{connectionId}/repositories/{owner}/{repo}/contents:
    get:
      description: Get files and directories at a specific path via OAuth connection
      parameters:
      - description: OAuth connection ID
        in: path
        name: connectionId
        required: true
        type: string
      - description: Repository owner
        in: path
        name: owner
        required: true
        type: string
      - description: Repository name
        in: path
        name: repo
        required: true
        type: string
      - description: Directory path (empty for root)
        in: query
        name: path
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get repository contents
      tags:
      - OAuth
...
  /auth/oauth/connections/{provider}:
    delete:
      description: Remove an OAuth connection for the authenticated user
      parameters:
      - description: Provider name (github, gitlab)
        in: path
        name: provider
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid provider
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Disconnect OAuth provider
      tags:
      - OAuth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchanges a refresh token for a new access token and refresh token
        pair
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth_interfaces_api_handlers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth_interfaces_api_handlers.RefreshTokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - auth
...
  /repositories:
    get:
      description: Retrieves a list of all repositories registered in OpsCore
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved repositories
          schema:
            $ref: '#/definitions/opscore_backend_internal_git_repository_interfaces_api_schema.ListRepositoriesResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/opscore_backend_internal_git_repository_interfaces_api_schema.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all repositories
      tags:
      - repositories
...
securityDefinitions:
  BearerAuth:
    description: 'Enter "Bearer " followed by your JWT token. Example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"